+++ test/wal-terminal-test.el
@@ -16,30 +16,41 @@
 
     (bydi-was-called eshell-truncate-buffer)))
 
-(ert-deftest wal-project-vterm ()
-  (bydi ((:always project-current)
+(ert-deftest wal-vterm--prefer-project ()
+  :tags '(terminal)
+
+  (defun faketerm (&rest args)
+    t)
+
+  (bydi ((:sometimes project-current)
          (:mock project-root :return "/tmp")
          (:mock project-prefixed-buffer-name :return "*test-vterm*")
          (:spy pop-to-buffer)
-         (:watch default-directory)
-         vterm)
+         (:spy faketerm)
+         (:watch default-directory))
 
-    (wal-project-vterm)
+    (wal-vterm--prefer-project 'faketerm t)
 
-    (bydi-was-called vterm)
+    (bydi-was-called faketerm)
     (bydi-was-set-to default-directory "/tmp")
 
     (ert-with-test-buffer (:name "vterm")
       (rename-buffer "*test-vterm*")
 
-      (funcall-interactively 'wal-project-vterm t)
+      (funcall-interactively 'wal-vterm--prefer-project 'faketerm t)
 
-      (bydi-was-called vterm)
+      (bydi-was-called faketerm)
       (bydi-was-not-called pop-to-buffer)
 
-      (wal-project-vterm)
+      (wal-vterm--prefer-project 'faketerm)
 
-      (bydi-was-called pop-to-buffer))))
+      (bydi-was-called pop-to-buffer))
+
+    (bydi-toggle-sometimes)
+
+    (funcall-interactively 'wal-vterm--prefer-project 'faketerm 'argument)
+
+    (bydi-was-called-last-with faketerm 'argument)))
 
 ;;; wal-terminal-test.el ends here